structural refactor of data classes
